refactor: Fix circular dependency between models
- Created models/database.py to centralize db instance
- Updated all imports to use models.database.db
- Removed db instantiation from user.py
- Fixed imports in 10+ files

Resolves circular dependency issue identified in code review
