+++ src/database/init_db.py
@@ -11,7 +11,8 @@ from datetime import datetime
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
 from main import app
-from models.user import db, User
+from models.database import db
+from models.user import User
 from models.prescription import (
     Prescription, Medication, ValidationResult, AuditLog,
     ValidationStatus, ProcessingStatus, InputFormat
