+++ src/main_integrated.py
@@ -69,7 +69,8 @@ except ImportError as e:
     admin_bp = None
 
 # Models
-from models.user import db, User
+from models.database import db
+from models.user import User
 from models.prescription import Prescription
 
 # ============================================================================
