+++ src/routes/user.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, jsonify, request
-from src.models.user import User, db
+from src.models.user import User
+from models.database import db
 
 user_bp = Blueprint('user', __name__)
